+++ NEWS.md
@@ -1,3 +1,9 @@
-# tapioca 0.0.2
+# tapioca 0.0.4 (7 December 2021)
+
+* Introduced `possibly` adverb for `doogal_download` and when properly documented, fixes issue #5.
+
+# tapioca 0.0.3 (7 December 2021)
 
 * Added a `NEWS.md` file to track changes to the package.
+* Added Counties API (doogal.co.uk) for retrieval of postcodes in `doogal_download.R` (addresses #1)
+* Improved some function param documentation